refactor(db): move utils.DataSize to db pkg
