+++ db/size.go
@@ -1,7 +1,14 @@
-package utils
+package db
 
 import (
 	"fmt"
+)
+
+const (
+	Kilobyte = 1024
+	Megabyte = 1024 * Kilobyte
+	Gigabyte = 1024 * Megabyte
+	Terabyte = 1024 * Gigabyte
 )
 
 type DataSize float64
