--- utils/const.go
@@ -1,8 +1,0 @@
-package utils
-
-const (
-	Kilobyte = 1024
-	Megabyte = 1024 * Kilobyte
-	Gigabyte = 1024 * Megabyte
-	Terabyte = 1024 * Gigabyte
-)
